better approach to determine if a user can be removed
The previous implementation of `canMemberBeRemoved` was inefficient as it                                                                                           │
fetched all groups and all their members to determine if a user was a                                                                                               │
member of any group.                                                                                                                                                │
                                                                                                                                                                    │
This commit refactors `canMemberBeRemoved` to be more performant. It                                                                                                │
now iterates through the groups and checks for the user's membership in                                                                                             │
each group, exiting as soon as a match is found. This avoids fetching                                                                                               │
all members of all groups, significantly reducing the number of API                                                                                                 │
calls and the amount of data processed.